+++ packages/components/src/components/animated-number/AnimatedNumberGroup.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+import { NumberFlowGroup } from '@number-flow/vue'
+</script>
+
+<template>
+  <NumberFlowGroup>
+    <slot />
+  </NumberFlowGroup>
+</template>