+++ packages/components/src/components/input-field-label/inputFieldLabelStyle.config.ts
@@ -1,6 +1,7 @@
 import { defineStyleConfig } from '@/styling/defineStyleConfig'
 
 export interface InputFieldLabelStyleConfig {
+  '--input-field-label-asterisk-color-default': string
   '--input-field-label-font-size-default': string
   '--input-field-label-font-weight-default': string
   '--input-field-label-spacing-default': string
@@ -11,6 +12,7 @@ export function setupDefaultInputFieldLabelStyles(): void {
   defineStyleConfig({
     colorScheme: '*',
     config: {
+      '--input-field-label-asterisk-color-default': 'var(--text-error-primary)',
       '--input-field-label-font-size-default': 'var(--text-sm)',
       '--input-field-label-font-weight-default': '500',
       '--input-field-label-spacing-default': 'var(--spacing-sm)',